Milestone 1
Not much going on in the software, mainly worked on getting a prototype LED matrix to be addressable and figuring out how to set up the ports so that they can address the matrix.

Getting the Atmega1284 to act as a power source was trivial, but there was some trouble getting the Atmega1284 to act as a ground. I figured out that the pins needed to be set as output and set to 0 to act as ground.